+++ src/app/lab/@tabs/layout/page.js
@@ -5,6 +5,7 @@ import Container from '@/display/Container' ;
 
 import LayoutDemo          from '@/demo/layouts/LayoutDemo' ;
 import DisplayDropdownDemo from '@/demo/layouts/DisplayDropdownDemo';
+import NavbarDemo          from '@/demo/layouts/NavbarDemo' ;
 
 /**
  * Layout showcase page.
@@ -22,6 +23,8 @@ const LayoutPage = () =>
         <LayoutDemo />
 
         <DisplayDropdownDemo />
+
+        <NavbarDemo />
     </Page>
 ) ;